+++ test/highlight_text_editor_test.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:highlight_text_editor/highlight_text_editor.dart';
-
-void main() {
-
-}